+++ cypress/plugins/index.js
@@ -1,21 +1,10 @@
 const axios = require('axios')
+const { sendSlackMessage } = require('../lib/slack')
+
 require('dotenv').config()
 
 module.exports = async (on, config) => {
-  on('after:run', async results => {
-    if (results) {
-      try {
-        const url = process.env.SLACK_WEBHOOK_URL
-        if (!url) {
-          return
-        }
-
-        await axios.post(process.env.SLACK_WEBHOOK_URL, buildSlackMessage(results))
-      } catch (error) {
-        console.error(error)
-      }
-    }
-  })
+  on('after:run', sendSlackMessage)
 
   let safeAppsList
 
@@ -35,64 +24,4 @@ module.exports = async (on, config) => {
   config.env.TESTING_SAFE_ADDRESS = process.env.CYPRESS_TESTING_SAFE_ADDRESS
 
   return config
-}
-
-const buildSlackMessage = results => {
-  const failedTests = results.runs[0].tests
-    .filter(test => test.state === 'failed')
-    .map(test => test.title[1])
-
-  const title = {
-    type: 'section',
-    text: {
-      type: 'mrkdwn',
-      text: '*Safe Apps e2e results*',
-    },
-  }
-
-  const executionResult = {
-    type: 'section',
-    text: {
-      type: 'mrkdwn',
-      text: `${results.totalPassed} out of ${results.totalTests}, passed`,
-    },
-  }
-
-  const failingApps = {
-    type: 'section',
-    text: {
-      type: 'mrkdwn',
-      text: `*Failing Apps:* _${failedTests.toString()}_`,
-    },
-  }
-
-  const action = {
-    type: 'section',
-    text: {
-      type: 'mrkdwn',
-      text: 'Want to take a look to the execution ?',
-    },
-    accessory: {
-      type: 'button',
-      text: {
-        type: 'plain_text',
-        text: 'Take me there !',
-        emoji: true,
-      },
-      value: 'click_me_123',
-      url: 'https://github.com/safe-global/safe-react-apps/actions',
-      action_id: 'button-action',
-    },
-  }
-
-  const blocks = [title, executionResult]
-
-  if (failedTests.length > 0) {
-    blocks.push(failingApps)
-  }
-  blocks.push(action)
-
-  return {
-    blocks,
-  }
 }
