+++ cypress/support/index.js
@@ -1,1 +1,1 @@
-import './commands'
+import './iframe'
